ntoskrnl.exe: Implement KeReleaseMutex() and waiting on mutexes.
We can get away with storing the mutex object in the KMUTEX for as long as
it's held, since on Windows it's illegal to destroy a mutex until it's
released.

Setting the flag OBJ_KERNEL_HANDLE on okfile is necessary to allow it to be
written to by system threads.