quartz/videorenderer: Get rid of hEvent.
The purpose of this event seems to be to ensure the filter is actually
flushed by the time EndFlush() returns. However, this is done quite badly
(e.g. we set hEvent before actually rendering the sample), and it's not clear
that it's necessary—the documentation implies that the upstream filter must
itself ensure that samples are flushed (in fact, before EndFlush() is called
on the downstream filter), and if this is done, there's no reason for us to
care.