d3dcompiler: Unconditionally report special (pixel shader output) sysvals based on the semantic name.
That is, do not check that the shader is a pixel shader and that the signature
is an output signature. These semantic names can only be generated in such
conditions; using the name for any other stage is invalid.

The motivation for this patch is that the correct sysval should be reported
even if the RDEF chunk is missing. Accordingly I have replaced each test
shader with a stripped version.