server: Use FileNameInformation to return device file object names.
This patch doesn't handle ntoskrnl device files. It is written with the ntoskrnl
tests previously written in mind, and therefore is structured to make
implementation somewhat easier, but the implementation is rather nontrivial and
so far not known to be necessary.

The equivalent Staging patch uses the object tree for device files. At least two
problems immediately arise from this: firstly, there is no clean way to
distinguish the root file name '\\' from the directly opened device (a
distinction particularly important for file systems, but in fact even visible
with named pipes); secondly, this cannot accomodate an implementation that must
plumb requests through ntoskrnl, since object names are set at creation time. In
general it also smacks of hacking a name into the object tree, since that name
is not actually looked up directly as other names are.

It may be preferable to separate retrieval of the parent into a separate fd_ops
callback.

Another possible design would be to leave the server implementation of
get_file_info alone, but alter get_object_info to return only the device name
for device file objects, and weld the two together in ntdll instead of in the
server.

The equivalent Staging patch was originally written for msys2's strace.exe.
msys2 and cygwin have been broken in Wine for a long time, perhaps unfixably.
I'm not aware of other applications that need pipes, or similar objects, to
return accurate names from ObjectNameInformation. I'm submitting this patch
partly to upstream a long-standing Staging patch and fix tests, and partly
because I believe it will help implementation of object/file name information
for regular files with some of my local/planned changes to volume
infrastructure.